pyproject.toml
[tool.poetry]
name = "beatcovidbackend"
version = "0.1.0"
description = ""
authors = ["beatcovi19now <git@beatcovid19now.org>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.3"
django-environ = "^0.4.5"
python-dotenv = "^0.12.0"
django-extensions = "^2.2.8"
djangorestframework = "^3.11.0"
coreapi = "^2.3.3"
markdown = "^3.2.1"
daphne = "^2.4.1"
sentry-sdk = "0.14.2"
django-cors-headers = "^3.2.1"
boto3 = "^1.12.16"
django-storages = "^1.9.1"
django-rest-auth = "^0.9.5"
huey = "^2.1"
redis = "*"
django-countries = "^6.1"
psycopg2-binary = "^2.8.4"
requests = "^2.23.0"

[tool.poetry.dev-dependencies]
black = "19.10b0"
flake8 = "^3.7.8"
isort = "^4.3.21"
pylint = "^2.2"
rope = "^0.16.0"
pylint-django = "^2.0.14"
ipython = "^7.13.0"
jupyter = "^1.0.0"

[tool.black]
line-length = 90
target_version = ['py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(\.git/
  |\.eggs
  |__pycache__
  |\.cache
  |\.ipynb_checkpoints
  |\.mypy_cache
  |\.pytest_cache
  |\.tox
  |\.venv
  |node_modules
  |_build
  |buck-out
  |build
  |dist
  |media
  |infrastructure
  |templates
  |locale
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 120

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
